Github OAuth demo with passport
