--- week6/temp/auth-example/ui/src/Dashboard.jsx
@@ -1,21 +1,0 @@
-import React from "react";
-import { useNavigate } from "react-router-dom";
-
-const Dashboard = () => {
-  const navigate = useNavigate();
-
-  const handleLogout = () => {
-    console.log("Logging out");
-    localStorage.removeItem("token");
-    navigate("/login");
-  };
-
-  return (
-    <div>
-      <h2>Dashboard</h2>
-      <button onClick={handleLogout}>Log out</button>
-    </div>
-  );
-};
-
-export default Dashboard;
