--- week6/w6d4/oauth-demo/frontend/src/App.jsx
@@ -1,19 +1,0 @@
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
-import './App.css'
-import Login from './Login'
-import Dashboard from './Dashboard'
-
-function App() {
-
-
-  return (
-    <Router>
-      <Routes>
-        <Route path="/login" element={<Login />} />
-        <Route path="/dashboard" element={<Dashboard />} />        
-      </Routes>
-    </Router>
-  )
-}
-
-export default App;
