--- week6/w6d4/oauth-demo/frontend/vite.config.js
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-})
