+++ week6/w6d4/oauth-with-passport/auth-example/ui/src/Dashboard.jsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import axios from 'axios';
+import { useState } from 'react'
+
+const Dashboard = () => {
+  const navigate = useNavigate();
+  const [userInfo, setUserInfo] = useState("");
+
+  const handleLogout = () => {
+    console.log("Logging out");
+    localStorage.removeItem("token");
+    navigate("/login");
+  };
+
+  const getResource = async () => {
+    const resp = await axios.get("http://localhost:3000/protected_route", {
+      headers: {
+        Authorization: `Bearer ${localStorage.getItem('token')}`
+      }
+    })
+    setUserInfo(resp.data.user.username);
+  }
+
+  return (
+    <div>
+      <h2>Dashboard</h2>
+      <button onClick={getResource}>Get Resource</button> 
+      <div>User name: {userInfo}</div>    
+      <button onClick={handleLogout}>Log out</button>
+    </div>
+  );
+};
+
+export default Dashboard;
