+++ week6/w6d4/oauth-with-passport/auth-example/ui/src/Login.jsx
@@ -1,8 +1,6 @@
 import React from "react";
-import { useNavigate } from "react-router-dom";
 
 const Login = () => {
-  const navigate = useNavigate();
 
   const handleLogin = () => {
     // This needs to be a full page reload to navigate to the GitHub OAuth page
